--- src/main/kotlin/com/devfloor/untitled/Home/domain/HomeBoardRepository.kt
@@ -1,8 +1,0 @@
-package com.devfloor.untitled.Home.domain
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.transaction.annotation.Transactional
-
-@Transactional(readOnly = true)
-interface HomeBoardRepository : JpaRepository<HomeBoard, Long> {
-}
