+++ src/main/kotlin/com/devfloor/untitled/board/domain/BoardRepository.kt
@@ -4,4 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.transaction.annotation.Transactional
 
 @Transactional(readOnly = true)
-interface BoardRepository : JpaRepository<Board, Long>
+interface BoardRepository : JpaRepository<Board, Long> {
+    fun findByTypeNot(boardType: BoardType): List<Board>
+}
